+++ src/NServiceBus.StructureMap.AcceptanceTests/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NServiceBus" version="6.0.0-unstable1187" targetFramework="net452" />
-  <package id="NServiceBus.AcceptanceTesting" version="6.0.0-unstable1187" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1616" targetFramework="net452" />
+  <package id="NServiceBus.AcceptanceTesting" version="6.0.0-unstable1616" targetFramework="net452" />
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
 </packages>
