--- src/NServiceBus.StructureMap.Tests/App_Packages/NSB.ContainerTests.6.0.0-unstable1187/TestContainerBuilder.cs
@@ -1,11 +1,0 @@
-namespace NServiceBus.ContainerTests
-{
-    using System;
-    using NServiceBus.ObjectBuilder.Common;
-
-    public static class TestContainerBuilder
-    {
-        public static Func<IContainer> ConstructBuilder = () => (IContainer)Activator.CreateInstance(Type.GetType("NServiceBus.ObjectBuilder.Autofac.AutofacObjectBuilder,NServiceBus.Core"));
-
-    }
-}
