+++ src/NServiceBus.StructureMap.Tests/App_Packages/NSB.ContainerTests.6.0.0-unstable1614/TestContainerBuilder.cs
@@ -1,0 +1,11 @@
+namespace NServiceBus.ContainerTests
+{
+    using System;
+    using NServiceBus.ObjectBuilder.Common;
+
+    public static class TestContainerBuilder
+    {
+        public static Func<IContainer> ConstructBuilder = () => (IContainer)Activator.CreateInstance(Type.GetType("NServiceBus.AutofacObjectBuilder,NServiceBus.Core"));
+
+    }
+}
