+++ src/NServiceBus.StructureMap.Tests/App_Packages/NSB.ContainerTests.6.0.0-unstable1614/When_disposing_the_builder.cs
@@ -35,28 +35,6 @@ namespace NServiceBus.ContainerTests
             builder.Dispose();
         }
 
-        [Test]
-        public void Should_dispose_all_IDisposable_components_in_child_container()
-        {
-            using (var main = TestContainerBuilder.ConstructBuilder())
-            {
-                DisposableComponent.DisposeCalled = false;
-                AnotherSingletonComponent.DisposeCalled = false;
-
-                main.RegisterSingleton(typeof(AnotherSingletonComponent), new AnotherSingletonComponent());
-                main.Configure(typeof(DisposableComponent), DependencyLifecycle.InstancePerUnitOfWork);
-
-                using (var builder = main.BuildChildContainer())
-                {
-                    builder.Build(typeof(DisposableComponent));
-                }
-                Assert.False(AnotherSingletonComponent.DisposeCalled, "Dispose should not be called on AnotherSingletonComponent because it belongs to main container");
-                Assert.True(DisposableComponent.DisposeCalled, "Dispose should be called on DisposableComponent");
-            }
-
-            //Not supported by, typeof(SpringObjectBuilder));
-        }
-
         public class DisposableComponent : IDisposable
         {
             public static bool DisposeCalled;
