+++ src/NServiceBus.StructureMap.Tests/App_Packages/NSB.ContainerTests.6.0.0-unstable1614/When_querying_for_registered_components.cs
