+++ src/NServiceBus.StructureMap.Tests/NServiceBus.StructureMap.Tests.csproj
@@ -36,7 +36,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="NServiceBus.Core, Version=6.0.0.0, Culture=neutral, PublicKeyToken=9fc386479f8a226c, processorArchitecture=MSIL">
-      <HintPath>..\packages\NServiceBus.6.0.0-unstable1187\lib\net452\NServiceBus.Core.dll</HintPath>
+      <HintPath>..\packages\NServiceBus.6.0.0-unstable1616\lib\net452\NServiceBus.Core.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="nunit.framework, Version=2.6.4.14350, Culture=neutral, PublicKeyToken=96d09a1eb7f44a77, processorArchitecture=MSIL">
@@ -52,13 +52,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1187\When_building_components.cs" />
-    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1187\When_registering_components.cs" />
-    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1187\When_using_nested_containers.cs" />
-    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1187\TestContainerBuilder.cs" />
-    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1187\When_disposing_the_builder.cs" />
-    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1187\When_querying_for_registered_components.cs" />
-    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1187\When_releasing_components.cs" />
+    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1614\TestContainerBuilder.cs" />
+    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1614\When_building_components.cs" />
+    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1614\When_disposing_the_builder.cs" />
+    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1614\When_querying_for_registered_components.cs" />
+    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1614\When_registering_components.cs" />
+    <Compile Include="App_Packages\NSB.ContainerTests.6.0.0-unstable1614\When_using_nested_containers.cs" />
     <Compile Include="SetUpFixture.cs" />
     <Compile Include="ShouldNotDisposeTheUnicastBus.cs" />
   </ItemGroup>
