+++ src/NServiceBus.StructureMap/packages.config
@@ -3,7 +3,7 @@
   <package id="Fody" version="1.29.3" targetFramework="net45" developmentDependency="true" />
   <package id="GitVersionTask" version="2.0.1" targetFramework="net452" developmentDependency="true" />
   <package id="Janitor.Fody" version="1.1.4.0" targetFramework="net452" developmentDependency="true" />
-  <package id="NServiceBus" version="6.0.0-unstable1187" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1616" targetFramework="net452" />
   <package id="NuGetPackager" version="0.5.5" targetFramework="net45" developmentDependency="true" />
   <package id="Particular.CodeRules" version="0.1.0" targetFramework="net452" developmentDependency="true" />
   <package id="structuremap" version="4.0.0.315" targetFramework="net452" />
